SUMMARY:  Changes to elminate compiler warnings
     BUGZILLA BUG #(s):  None.

     CHECKLIST
     ---------

     1) Did you 'cvs add' all new files and directories [Y:N:N/A]?  N

     2) Did you add new source files to a CMakeLists.txt file [Y:N:N/A]?  N

     3)  Did you modify or add any source or build system files?  N


     DESCRIPTION OF CHANGES:

     There are several compiler warnings from the Windows builds that have to
do with casting types (double to int, etc.).  I've added explicit casts to
take care of these warnings.  I've also initialized some variables to take
care of other warnings.

(checklist version 2009-07-14)

+++ packages/amesos/src/SuiteSparse/CAMD/Source/amesos_camd_2.c
@@ -725,7 +725,7 @@ GLOBAL void CAMD_2
     }
     else
     {
-	dense = alpha * sqrt ((double) n) ;
+	dense = (int) ( alpha * sqrt ((double) n) ) ;
     }
     dense = MAX (16, dense) ;
     dense = MIN (n,  dense) ;
